requirements.txt
absl-py==0.10.0
aiohttp==3.6.2
altgraph==0.17
astunparse==1.6.3
async-timeout==3.0.1
attrs==20.2.0
backcall==0.2.0
beautifulsoup4==4.9.3
bokeh==2.2.3
boto3==1.16.9
botocore==1.19.9
bottle==0.12.18
bottle-websocket==0.2.9
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
cloudipsp==1.0.2
colorama==0.4.3
configparser==5.0.0
decorator==4.4.2
dnspython==2.0.0
dominate==2.6.0
Eel==0.14.0
filelock==3.0.12
Flask==1.1.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.5.0
Flask-MySQLdb==0.2.0
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
future==0.18.2
fxcmpy==1.2.6
gast==0.3.3
gevent==20.9.0
gevent-websocket==0.10.1
google-auth==1.22.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
greenlet==0.4.17
grpcio==1.32.0
h5py==2.10.0
idna==2.10
ipython==7.18.1
ipython-genutils==0.2.0
is-disposable-email==1.0.0
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.17.0
Keras-Preprocessing==1.1.2
Markdown==3.2.2
MarkupSafe==1.1.1
multidict==4.7.6
mysql-connector==2.2.9
mysql-connector-python==8.0.22
mysqlclient==2.0.1
numpy==1.18.5
oauthlib==3.1.0
opencv-python==4.4.0.46
opt-einsum==3.3.0
packaging==20.4
pandas==1.1.2
panel==0.10.1
param==1.10.0
parso==0.7.1
pefile==2019.4.18
pickleshare==0.7.5
Pillow==8.0.1
prompt-toolkit==3.0.7
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyct==0.4.8
Pygments==2.7.1
pyinstaller==4.0
pyinstaller-hooks-contrib==2020.9
pyparsing==2.4.7
PyPDF2==1.26.0
pypng==0.0.20
PyQRCode==1.2.1
PySide2==5.15.1
PySocks==1.7.1
python-dateutil==2.8.1
python-engineio==3.13.2
python-socketio==4.6.0
pytorch-transformers==1.2.0
pytz==2020.1
pyviz-comms==0.7.6
pywin32-ctypes==0.2.0
PyYAML==5.3.1
regex==2020.10.28
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
s3transfer==0.3.3
sacremoses==0.0.43
sentencepiece==0.1.94
shiboken2==5.15.1
six==1.15.0
socketIO-client==0.7.2
soupsieve==2.0.1
SQLAlchemy==1.3.19
StackAPI==0.1.12
tensorboard==2.1.1
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.1
tensorflow-estimator==2.3.0
tensorflow-gpu==2.3.1
tensorflow-gpu-estimator==2.2.0
termcolor==1.1.0
tokenizers==0.9.2
torch @ file:///C:/Users/Cavad/Downloads/torch-1.6.0-cp38-cp38-win_amd64.whl
tornado==6.1
tqdm==4.51.0
traitlets==5.0.4
transformers==3.4.0
typing-extensions==3.7.4.3
urllib3==1.25.10
visitor==0.1.3
wcwidth==0.2.5
websocket-client==0.57.0
Werkzeug==1.0.1
whichcraft==0.6.1
wikipedia==1.4.0
wrapt==1.12.1
WTForms==2.3.3
wtforms-validators==1.0.0
wxPython==4.1.0
yarl==1.6.0
zope.event==4.5.0
zope.interface==5.1.2
